Many To Many Commit

+++ src/main/java/com/example/demo/repository/CoursesRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CoursesRepository extends JpaRepository<Courses, Integer>
+public interface CoursesRepository extends JpaRepository<Courses, Long>
 {
 
 }
